% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{CalcuRsquare}
\alias{CalcuRsquare}
\title{CalcuRsquare}
\usage{
CalcuRsquare(y, y.pred)
}
\arguments{
\item{y}{numeric}

\item{y.pred}{numeric}
}
\description{
calcualte r2 for regression
}
\author{
Zhiwei Zhou
}
