% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CCS_Calu_SingleField.R
\name{CCS_Calu_SingleField}
\alias{CCS_Calu_SingleField}
\title{CCS_Calu_SingleField}
\usage{
\code{CCS_Calu_SingleField(polarity="pos", charge=1, Agilent.table="V1")}
}
\arguments{
\item{raw.data}{Character. The excel file name. The top 3 columns should be named as: 'M', 'Adducts', 'Arrival.time'}

\item{polarity}{Character. Ionzation polarity of experiments. Should be one of "pos" or "neg". "pos" denotes "positive mode", "neg" denotes "negative mode". Default is "pos".}

\item{charge}{Numeric. The charge of calibrants. Default is 1.}

\item{Agilent.table}{Character. The verison of Agilent tuning mix reference table is used to establish calibration Scurve. "V1" denotes the reference table in Agilent IM-MS browser B.07.01; "V2" denotes the reference table reported by Stow et. al. Anal. Chem. 2017}
}
\value{
a csv table includs, m/z and CCS values etc.
}
\description{
Calculate CCS values using single field method
}
\examples{
CCS_Calu_SingleField(polarity="pos", charge=1, Agilent.table="V1")

}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
