% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalCalculation.R
\name{CiCalcu}
\alias{CiCalcu}
\title{CiCalcu}
\usage{
CiCalcu(x, cred_para_set, q = 0.995)
}
\arguments{
\item{x}{Power fitting model (Generated by FitPowerFunction)}

\item{cred_para_set}{Power fitting result (Generated by FitPowerFunction)}

\item{q}{vector of quantiles; Default: 0.995, representing 99% cridential}
}
\description{
Calculate confidence interval (CI) with input of x
}
\author{
Zhiwei Zhou
}
