% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalCalculation.R
\name{CredParaSetCalcu}
\alias{CredParaSetCalcu}
\title{CredParaSetCalcu}
\usage{
CredParaSetCalcu(fit_model, fit_result)
}
\arguments{
\item{fit_model}{Power fitting model (Generated by FitPowerFunction)}

\item{fit_result}{Power fitting result (Generated by FitPowerFunction)}
}
\description{
Calculate parameter set for CI and PI calucaltion
}
\author{
Zhiwei Zhou
}
