% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCfmid.R
\name{extractCfmid}
\alias{extractCfmid}
\title{extractCfmid}
\usage{
extractCfmid(file_name, mode = c("spec_only", "spec_annotate"))
}
\arguments{
\item{file_name}{the name of log files. e.g. C00182.log}

\item{mode}{'spec_only' or 'spec_annotate'. spec_only: extract spec; spec_annotate: extract spec with annotate}
}
\description{
extract cfm_id predicted spec, .log
}
\examples{
extractCfmid("C00182.log", mode = 'spec_annotate')
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
