% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractClassyFire.R
\name{extractClassyFire}
\alias{extractClassyFire}
\title{extractClyssyFire}
\usage{
extractClassyFire(filename)
}
\arguments{
\item{filename}{the sdf file name}
}
\description{
extract sdf file exported from ClassyFire
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
