% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExtractTargetMsMsSpectra.R
\name{ExtractTargetMsMS}
\alias{ExtractTargetMsMS}
\title{ExtractTargetMsMs}
\usage{
ExtractTargetMsMS(msms_file, polarity = c("positive", "negative"),
  mode = c("SPMCE", "MPSCE", "SPSCE"), ce_levels = c("CE10", "CE20",
  "CE40"))
}
\arguments{
\item{msms_file}{mzXML file format}

\item{polarity}{'positive', 'negative'. Default: 'positive'}

\item{mode}{SPMCE: single precursor with multiple CEs; MPSCE: multiple precursors, but single CE for each precursor; SPSCE: single precursor with one CE;}

\item{ce_levels}{collison energy. Default: c("CE10", "CE20", "CE40")}
}
\description{
ExtractTargetMsMs
}
\examples{
setwd('F:/01 MetIMMS/00 data processing/190702 extract msms spectra from targeted 6560 data/convert_msms')
test <- ExtractTargetMsMS(msms_file = "L0656_H_pos_10ppm.mzXML", polarity = 'positive', mode = 'SPMCE', ce_levels = c("CE10", "CE20", "CE40"))
}
\author{
Zhiwei Zhou
}
