% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IntervalCalculation.R
\name{FitPowerFunction}
\alias{FitPowerFunction}
\title{FitPowerFunction}
\usage{
FitPowerFunction(raw_data, id = "id", x = "x", y = "y")
}
\arguments{
\item{raw_data}{data.frame, it need 3 columns at least}

\item{id}{character. column name of id}

\item{x}{character. column name of x}

\item{y}{character. column name of y}
}
\description{
Power fitting
}
\author{
Zhiwei Zhou
}
