% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsilicoMSMSFunction.R
\name{GenerateCfmIdCandidates}
\alias{GenerateCfmIdCandidates}
\title{GenerateCfmIdCandidates}
\usage{
GenerateCfmIdCandidates(metfrag_candidate)
}
\arguments{
\item{metfrag_candidate}{the result generated from MetFrag}
}
\description{
generate candidate list for CFM-ID
}
\author{
Zhiwei Zhou
}
