% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsilicoMSMSFunction.R
\name{GenerateCfmIdSpec}
\alias{GenerateCfmIdSpec}
\title{GenerateCfmIdSpec}
\usage{
GenerateCfmIdSpec(energy0 = NULL, energy1 = NULL, energy2 = NULL,
  dir_path = ".", file_name = "temp_spec.txt")
}
\arguments{
\item{energy0}{10v specttra.  1st column: mz; 2nd column: intensity.}

\item{energy1}{20v specttra.  1st column: mz; 2nd column: intensity.}

\item{energy2}{40v specttra.  1st column: mz; 2nd column: intensity.}

\item{dir_path}{Default: '.'}

\item{file_name}{Default:'temp_spec.txt'}
}
\description{
generate peak list for CFM-ID
}
\examples{
GenerateCfmIdSpec(energy)
}
\author{
Zhiwei Zhou
}
