% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsilicoMSMSFunction.R
\name{GenerateMetFragLocalDB}
\alias{GenerateMetFragLocalDB}
\title{GenerateMetFragLocalDB}
\usage{
GenerateMetFragLocalDB(metfrag_candidate)
}
\arguments{
\item{metfrag_candidate}{the result generated from MetFrag}
}
\description{
generate candidate list for MetFrag
}
\author{
Zhiwei Zhou
}
