% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryTool.R
\name{GetMsMs}
\alias{GetMsMs}
\title{GetMsMs}
\usage{
GetMsMs(database, polarity = c("pos", "neg"), id = "L0002",
  instrument = c("ABSciex", "Agilent"), ce = c("10", "15", "20", "25",
  "30", "35", "35,15", "40", "45", "50", "55", "60", "65", "70"))
}
\arguments{
\item{database}{a database with zhulib info}

\item{polarity}{'pos', 'neg'. Default: 'pos'}

\item{id}{a character of lib id}

\item{instrument}{'ABSciex', 'Agilent'}

\item{ce}{character. '10', '15', '20', '25', '30', '35', '35,15', '40', '45', '50', '55', '60', '65', '70'}
}
\value{
a data.frame with compound info
}
\description{
extract compound MS/MS spectra from zhulib
}
\author{
Zhiwei Zhou
}
