% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryTool.R
\name{GetMsMsInfo}
\alias{GetMsMsInfo}
\title{GetMsMsInfo}
\usage{
GetMsMsInfo(database, id = "L0002")
}
\arguments{
\item{database}{a database with zhulib info}

\item{id}{a character of lib id}
}
\value{
a data.frame with compound info
}
\description{
extract compound information with zhulib
}
\author{
Zhiwei Zhou
}
