% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MdCleaning.R
\name{MdCleaning}
\alias{MdCleaning}
\title{MdCleaning}
\usage{
MdCleaning(raw.data, col.number.res)
}
\arguments{
\item{raw.data}{a data.frame of MD, each column represent one MD, and each row represent one compound}

\item{col.number.res}{the column index to reserve}
}
\description{
Molecular Descriptor Cleaning to remove NA and same molecular descriptors
}
\examples{
MdCleaning
}
\author{
Zhiwei Zhou
}
