% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsilicoMSMSFunction.R
\name{MetFragMatch}
\alias{MetFragMatch}
\title{MetFragMatch}
\usage{
MetFragMatch(DatabaseSearchRelativeMassDeviation = 5,
  FragmentPeakMatchAbsoluteMassDeviation = 0.001,
  FragmentPeakMatchRelativeMassDeviation = 5,
  MetFragDatabaseType = "PubChem", NeutralPrecursorMass = 253.966126,
  PeakList)
}
\arguments{
\item{DatabaseSearchRelativeMassDeviation}{Default: 5ppm}

\item{FragmentPeakMatchAbsoluteMassDeviation}{Default: 0.001Da}

\item{FragmentPeakMatchRelativeMassDeviation}{Default: 5.0ppm}

\item{MetFragDatabaseType}{'PubChem'}

\item{NeutralPrecursorMass}{253.966126}

\item{PeakList}{matrix with 2 columns, 1st: mz, 2nd: intensity}
}
\value{
a data.frame with scores and candidate properties
}
\description{
MetFragMatch
}
\author{
Zhiwei Zhou
}
