% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{msms_plot}
\alias{msms_plot}
\title{msms_plot}
\usage{
msms_plot(raw.exp.spec, raw.std.spec, match.result = NULL, dir.path = NULL,
  cmp.name = NULL, is.output = c(TRUE, FALSE), pg.formate = c("png",
  "pdf"))
}
\arguments{
\item{raw.exp.spec}{Matrix or data frame. The experiment MS/MS spectrum. The 1st column is "mz", the 2nd column is "intensity".}

\item{raw.std.spec}{Matrix or data frame. The library MS/MS spectrum. The 1st column is "mz", the 2nd column is "intensity"}

\item{match.result}{output from SpecMatch}

\item{dir.path}{the file path of output}

\item{cmp.name}{Character. The name of compound}

\item{is.output}{Logical. Whether output the plot?}

\item{pg.format}{png or pdf format.}
}
\description{
plot MS/MS mirror plot.
}
\examples{
raw.exp.spec <- data.frame(mz=c(42.0364, 58.0664, 59.0745, 118.0858), intensity=c(0.01, 1, 0.4, 0.25))
raw.std.spec <- data.frame(mz=c(42.0364, 58.0664, 75.4568, 118.0858), intensity=c(0.01, 1, 0.2, 0.25))
msms_plot(raw.exp.spec, raw.std.spec, is.output=FALSE)
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
