% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{readMSP}
\alias{readMSP}
\title{readMSP}
\usage{
readMSP(file)
}
\value{
A list object. info: MS1 information, including ms1, rt etc.; spec: the MS/MS spectrum
}
\description{
read MSP spectra files
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
