% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BioTransformerFunctions.R
\name{RunBioTransformer}
\alias{RunBioTransformer}
\title{RunBioTransformer}
\usage{
RunBioTransformer(pgm_path, smiles, output_file, task = c("pred", "cid"),
  bt_type = c("ecbased", "cyp450", "phaseII", "hgut", "superbio",
  "allHuman", "envimicro"), str_type = c("-ismi"),
  output_format = c("-ocsv"), n_step = 1)
}
\arguments{
\item{pgm_path}{the path of biotransformer program. Default: 'I:/software/Tools/BioTransformer-master'}

\item{smiles}{the smiles of structure}

\item{output_file}{the csv name of output}

\item{task}{'pred', 'cid'. prediction or compound identification. Default: pred}

\item{bt_type}{'ecbased', 'cyp450', 'phaseII', 'hgut', 'superbio', 'allHuman', 'envimicro'. Default: 'ecbased'. Ref: https://github.com/JustinZZW/BioTransformer}

\item{str_type}{Default: '-ismi'}

\item{output_format}{Default: '-ocsv'}

\item{n_step}{number for reaction. Default: 1}
}
\description{
R interface to run BioTransformer in CMD. Please refer Djoumbou-Feunang Y, Fiamoncini J, de la Fuente AG, Manach C, Greiner R, and Wishart DS; BioTransformer: A Comprehensive Computational Tool for Small Molecule Metabolism Prediction and Metabolite Identification; Journal of Cheminformatics 201911:2; DOI: 10.1186/s13321-018-0324-5.
}
\examples{
RunBioTransformer(pgm_path='I:/software/Tools/BioTransformer-master',
smiles = "CC(C)C1=CC=C(C)C=C1O",
output_file='I:/software/Tools/BioTransformer-master/test.csv',
n_step=1)
}
\author{
Zhiwei Zhou
}
