% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InsilicoMSMSFunction.R
\name{RunCFMID}
\alias{RunCFMID}
\title{RunCFMID}
\usage{
RunCFMID(cfm_id = "F:/MetIMMS_MetFrag/cfm_id/cfm-id.exe",
  spec_file = "F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_test_result/example_spec.txt",
  id = "temp",
  candidate_file = "F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_test_result/example_candidates.txt",
  num_highest = -1, ppm_mass_tol = 10, abs_mass_tol = 0.01,
  prob_thresh = 0.001,
  param_file = "F:/MetIMMS_MetFrag/cfm_id/metab_se_cfm/param_output0.log",
  config_file = "F:/MetIMMS_MetFrag/cfm_id/metab_se_cfm/param_config.txt",
  score_type = c("Jaccard", "DotProduct"), apply_postprocessing = 1,
  output_file = "F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_output/TEST_01_result.txt",
  output_msp = "F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_output/TEST_01_spec.msp")
}
\arguments{
\item{cfm_id}{the path of cfm_id.exe Default: 'F:/MetIMMS_MetFrag/cfm_id/cfm-id.exe'}

\item{spec_file}{the path of spec_file txt file. 1st column: mz; 2nd column: intensity. It could be generated by GenerateCfmIdSpec}

\item{id}{An identifier for the target molecule (Used to retrieve input spectrum from msp (if used). Default: 'temp'}

\item{num_highest}{The number of (ranked) candidates to return or -1 for all. Default: -1}

\item{ppm_mass_tol}{The mass tolerance in ppm to use when matching peaks within the dot product comparison - will use higher resulting tolerance of ppm and abs (if not given defaults to 10ppm). Default: 10}

\item{abs_mass_tol}{The mass tolerance in abs Da to use when matching peaks within the dot product comparison - will use higher resulting tolerance of ppm and abs ( if not given defaults to 0.01Da). Default: 0.01}

\item{prob_thresh}{The probability below which to prune unlikely fragmentations (default 0.001). Default: 0.001}

\item{param_file}{This file is the output of cfm-train. Pre-trained models as used in the above publication can be found in the supplementary data for that paper stored within the source tree of this project. Here, we use the se sfm model for both positive and negative modes. Positive mode: F:/MetIMMS_MetFrag/cfm_id/metab_se_cfm/param_output0.log; Negative mode: F:/MetIMMS_MetFrag/cfm_id/negative_metab_se_cfm/param_output0.log. Default:  'F:/MetIMMS_MetFrag/cfm_id/metab_se_cfm/param_output0.log'}

\item{config_file}{The filename where the configuration parameters of the cfm model can be found. This needs to match the file passed to cfm-train during training. See cfm-train documentation below for further details. Here, we use the se sfm model for both positive and negative modes. Positive mode: F:/MetIMMS_MetFrag/cfm_id/metab_se_cfm/param_config.txt; Negative mode: F:/MetIMMS_MetFrag/cfm_id/negative_metab_se_cfm/param_config.txt. Default:  'F:/MetIMMS_MetFrag/cfm_id/metab_se_cfm/param_config.txt'}

\item{score_type}{The type of scoring function to use when comparing spectra. Options: Jaccard (default), DotProduct. Default: Jaccard}

\item{apply_postprocessing}{Whether or not to post-process predicted spectra to take the top 80% of energy (at least 5 peaks), or the highest 30 peaks (whichever comes first) (0 = OFF (default for EI-MS), 1 = ON (default for ESI-MS/MS)). Default: 1}

\item{output_file}{the path of output result. It should be in ".txt" format. Default: 'F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_output/TEST_01_result.txt'}

\item{output_msp}{the path of output MSP file (MS/MS file). It should be in ".msp" format. Default: 'F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_output/TEST_01_spec.msp'}

\item{spec_file}{the path of spec_file txt file. 1st column: identifier; 2nd column: SMILES. It could be generated by GenerateCfmIdSpec}
}
\description{
R interfaces to run CFM-ID.exe. Detail: https://sourceforge.net/p/cfm-id/wiki/Home/#cfm-id-precomputed
}
\examples{
RunCFMID(spec_file = 'F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_test_result/example_spec.txt',
         candidate_file = 'F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_test_result/example_candidates.txt',
         ppm_mass_tol = 10,
         abs_mass_tol = 0.01,
         score_type = 'Jaccard',
         output_file = 'F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_output/TEST_01_result.txt',
         output_msp = 'F:/MetIMMS_MetFrag/20190521_retrieve_candidates/test/cfm_id_output/TEST_01_spec.msp')
}
\author{
Zhiwei Zhou
}
