% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LibraryTool.R
\name{SplitInchiKey}
\alias{SplitInchiKey}
\title{SplitInchiKey}
\usage{
SplitInchiKey(inchikey = "AEMRFAOFKBGASW-UHFFFAOYSA-N")
}
\value{
a data.frame with 4 columns
}
\description{
split inchikey to 3 parts
}
\author{
Zhiwei Zhou
}
