% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{ZZWImpute}
\alias{ZZWImpute}
\title{impute}
\usage{
ZZWImpute(raw_matrix, label, k = 10, rowmax = 0.5, colmax = 0.8,
  maxp = 1500, rng.seed = 362436069)
}
\arguments{
\item{raw_matrix}{the numeric matrix}

\item{label}{the label of each row}

\item{k}{please ref impute.knn. Default: 10}

\item{rowmax}{please ref impute.knn. Default: 0.5}

\item{colmax}{please ref impute.knn. Default: 0.8}

\item{maxp}{please ref impute.knn. Default: 1500}

\item{rng.seed}{please ref impute.knn. Default: 362436069}
}
\description{
impute
}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
