% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mz_transform}
\alias{mz_transform}
\title{mz_transform}
\usage{
\code{mz_transform(M, adduct, polarity=c('pos', 'neg'))}
}
\arguments{
\item{M}{a vector of exact mass.}

\item{adduct}{a vector of adducts. These adducts need to be}

\item{formula}{sdf}

\item{polarity}{'pos' or 'neg'}
}
\examples{
mz_transform(M = 123.2324, adduct = c('[M+H]+', '[M+Na]+'), polarity = 'pos')

}
\author{
Zhiwei Zhou
\email{zhouzw@sioc.ac.cn}
}
